move functions into their own files
